use MDX in multiple layouts
