+++ layouts/blog.js
@@ -1,0 +1,38 @@
+import { MDXProvider } from '@mdx-js/react';
+import dayjs from 'dayjs';
+import utc from 'dayjs/plugin/utc';
+import MDXComponents from '../lib/mdx-components';
+import Layout from '../components/layout';
+import Head from '../components/head';
+import Link from '../components/link';
+import { H1 } from '../components/headings';
+
+dayjs.extend(utc);
+
+export default function Index({ children, frontMatter }) {
+  const { title, date, tags, image, description } = frontMatter;
+
+  return (
+    <Layout>
+      <Head
+        title={title}
+        date={date}
+        description={description}
+        image={image}
+        tags={tags}
+        isArticle
+      />
+      <div>
+        <H1>{title}</H1>
+        <p className="mb-8">{dayjs.utc(date).format('MMMM D, YYYY')}</p>
+        <hr />
+        <MDXProvider components={MDXComponents}>{children}</MDXProvider>
+        <div className="flex items-center justify-center my-8">
+          <Link href="/blog">
+            &larr; Back to Blog
+          </Link>
+        </div>
+      </div>
+    </Layout>
+  );
+}
