+++ layouts/projects/default.js
@@ -1,5 +1,8 @@
+import { MDXProvider } from '@mdx-js/react';
 import Head from '../../components/head';
 import Layout from '../../components/layout';
+import { H1 } from '../../components/headings';
+import MDXComponents from '../../lib/mdx-components';
 
 export default function Index({ children, frontMatter }) {
   const { name, description, image, date } = frontMatter;
@@ -14,10 +17,8 @@ export default function Index({ children, frontMatter }) {
         isArticle
       />
       <div>
-        <h1>{name}</h1>
-        <div>
-          {children}
-        </div>
+        <H1>{name}</H1>
+        <MDXProvider components={MDXComponents}>{children}</MDXProvider>
       </div>
     </Layout>
   );
