+++ pages/404.md
@@ -1,0 +1,6 @@
+---
+title: 404
+---
+
+# 404
+The page you requested doesn't exist. [Return home](/)?
