+++ pages/blog/2020-01-06-migrating-my-website-to-aws-using-terraform.mdx
@@ -3,6 +3,7 @@ title:        Migrating My Website to AWS Using Terraform
 description:  I recently migrated my personal website infrastructure from a VPS to AWS S3 + CloudFront, with infrastructure managed by Terraform.
 image:        /img/waterfall-cdn.png
 date:         "2020-01-06T00:00:00Z"
+layout:       blog
 tags:
   - terraform
   - aws
