+++ pages/blog/2020-01-17-tufte.mdx
@@ -3,6 +3,7 @@ title:        "Edward Tufte: Presenting Data and Information"
 description:  Some learnings from Edward Tufte's "Presenting Data and Information" course
 image:        /img/tufte/tufte-sf.jpeg
 date:         "2020-01-17T00:00:00Z"
+layout:       blog
 tags:
   - data visualization
   - edward tufte
